requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
psycopg2-binary